🚀 Refactor some classes
